Mobed network connection check logic from GDPRConsentLib to SOurepoint class and webview class
