+++ cmplibrary/src/main/java/com/sourcepoint/gdpr_cmplibrary/SourcePointClient.java
@@ -1,5 +1,7 @@
 package com.sourcepoint.gdpr_cmplibrary;
 
+import android.net.ConnectivityManager;
+import android.net.NetworkInfo;
 import android.util.Log;
 
 import org.json.JSONException;
@@ -30,18 +32,32 @@ class SourcePointClient {
 
     SourcePointClientConfig config;
 
+    ConnectivityManager manager;
+
     private String getRequestUUID(){
         if(!requestUUID.isEmpty()) return requestUUID;
         requestUUID =  UUID.randomUUID().toString();
         return requestUUID;
     }
 
-    SourcePointClient(OkHttpClient httpClient, SourcePointClientConfig config) {
+    SourcePointClient(OkHttpClient httpClient, SourcePointClientConfig config, ConnectivityManager manager) {
         this.httpClient = httpClient;
         this.config = config;
+        this.manager = manager;
+    }
+
+    private boolean hasLostInternetConnection() {
+        if (this.manager == null) {
+            return true;
+        }
+        NetworkInfo activeNetwork = this.manager.getActiveNetworkInfo();
+        return activeNetwork == null || !activeNetwork.isConnectedOrConnecting();
     }
 
     void getMessage(boolean isNative, String consentUUID, String meta, String euconsent, GDPRConsentLib.OnLoadComplete onLoadComplete) throws ConsentLibException {
+        if(hasLostInternetConnection())
+            throw new ConsentLibException.NoInternetConnectionException();
+
         String url = messageUrl(isNative);
         Log.d(LOG_TAG, "Getting message from: " + url);
 
@@ -110,6 +126,9 @@ class SourcePointClient {
 
 
     void sendConsent(JSONObject params, GDPRConsentLib.OnLoadComplete onLoadComplete) throws ConsentLibException {
+        if (hasLostInternetConnection())
+            throw new ConsentLibException.NoInternetConnectionException();
+
         String url = consentUrl();
         try {
             params.put("requestUUID", getRequestUUID());
